More improvements.
Now if you have a has_one embedded object, even if it's not defined, you can access it using:
$object->embbed_has_one_object

This is useful if you specify default values in your embedded object, and only save data if the values are different from the defaults.

Even when the embedded object has not been set, you can change values using
$object->embedded->key = 'new_value'

It will update accordingly.